requirements.txt
python-libtorrent-bin==2.0.2
ujson==5.2.0
easydict==1.9